(joddev) release JAL release 0.9.4
fixed a nasty (make) bug that has lingered
unnoticed in JOD for years.  When words with
the same name and different name classes
exist in different dictionaries on the path
the make failed due to improper name classification
of words at depth greater than 1. It's surprising
I never noticed this bug because it's definitely
not the way I intended JOD to operate.