add (^:)=:  commet tag to override mixed assignments
The previous fix exposed a problem with JOD code. There are a number of
legitimate uses of mixed indirect assignments. This new comment tag
instructs (globs) to ignore mixed assignments. Usually a name that is
assigned both local and global indicates an error or poor name choice
but there are valuable uses for such a pattern and this exception allows
JOD to handle them.